small test refactor
